chore(ui): seprate routes as per categories
