--- openmetadata-ui/src/main/resources/ui/src/components/Entity/EntityDetails/EntityDetails.component.tsx
@@ -1,27 +1,0 @@
-/*
- *  Copyright 2024 Collate.
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *  http://www.apache.org/licenses/LICENSE-2.0
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-import React, { useMemo } from 'react';
-import { useParams } from 'react-router-dom';
-import { EntityType } from '../../../enums/entity.enum';
-import entityUtilClassBase from '../../../utils/EntityUtilClassBase';
-
-export const EntityDetail = () => {
-  const { entityType } = useParams<{ entityType: EntityType }>();
-
-  const Component = useMemo(
-    () => entityUtilClassBase.getEntityDetailComponent(entityType),
-    [entityType]
-  );
-
-  return <Component />;
-};
